+++ config/initializers/simple_form.rb
@@ -14,14 +14,10 @@ SimpleForm.setup do |config|
     b.use :input, class: 'govuk-input', error_class: 'govuk-input--error'
   end
 
-  config.wrappers :govuk_textarea_wrapper, tag: 'div', class: 'govuk-textarea-wrapper',
-error_class: 'govuk-error' do |b|
-    b.use :html5
-    b.use :placeholder
-    b.optional :maxlength
-    b.optional :minlength
-    b.use :error, wrap_with: { tag: 'span', class: 'govuk-error-message' }
-    b.use :input, class: 'govuk-textarea', error_class: 'govuk-input--error'
+  config.wrappers :govuk_textarea_wrapper, parent: :default do |b|
+    b.use :label, class: 'govuk-label'
+    b.use :full_error, wrap_with: { tag: 'span', class: 'govuk-error-message' }
+    b.use :input, class: 'govuk-textarea', input_html: { rows: 5 }
   end
 
   # The default wrapper to be used by the FormBuilder.