+++ lib/custom_markdown_renderer.rb
@@ -2,10 +2,10 @@ require 'redcarpet'
 
 class CustomMarkdownRenderer < Redcarpet::Render::HTML
   def link(link, title, content)
-    "<a href=\"#{link}\" title=\"#{title}\" class=\"govuk-notification-banner__link\" target=\"_blank\">#{content}</a>"
+    "<a href=\"#{link}\" title=\"#{title}\" class=\"govuk-link\" target=\"_blank\">#{content}</a>"
   end
 
   def paragraph(content)
-    "<p class=\"govuk-notification-banner__heading\">#{content}</p>"
+    "<p class=\"govuk-body\">#{content}</p>"
   end
 end